% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-ci-data.R
\name{addFactorByFactorDifferenceOfDifferencesRow}
\alias{addFactorByFactorDifferenceOfDifferencesRow}
\title{Add the two-by-two interaction contrast for a factor-by-factor model}
\usage{
addFactorByFactorDifferenceOfDifferencesRow(
  model,
  mf,
  level,
  firstFactor,
  secondFactor,
  appendRow
)
}
\arguments{
\item{model}{A fitted model object.}

\item{mf}{Model frame.}

\item{level}{Confidence level.}

\item{firstFactor}{First factor name.}

\item{secondFactor}{Second factor name.}

\item{appendRow}{Row appender closure.}
}
\value{
Invisibly returns \code{NULL}.
}
\description{
Add the two-by-two interaction contrast for a factor-by-factor model
}
\keyword{internal}
