% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-ci-data.R
\name{addLogisticFactorByFactorRatioOfOddsRatiosRow}
\alias{addLogisticFactorByFactorRatioOfOddsRatiosRow}
\title{Add the two-by-two ratio-of-odds-ratios row for a logistic interaction}
\usage{
addLogisticFactorByFactorRatioOfOddsRatiosRow(
  model,
  mf,
  level,
  firstFactor,
  secondFactor,
  appendRow
)
}
\arguments{
\item{model}{A fitted logistic model.}

\item{mf}{Model frame.}

\item{level}{Confidence level.}

\item{firstFactor}{First factor name.}

\item{secondFactor}{Second factor name.}

\item{appendRow}{Row appender closure.}
}
\value{
Invisibly returns \code{NULL}.
}
\description{
Add the two-by-two ratio-of-odds-ratios row for a logistic interaction
}
\keyword{internal}
