% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis-recipe-download.R
\name{analysisRecipeDownloadExtension}
\alias{analysisRecipeDownloadExtension}
\title{Return the download extension for an analysis recipe}
\usage{
analysisRecipeDownloadExtension(recipe)
}
\arguments{
\item{recipe}{A \code{wmfmAnalysisRecipe} object.}
}
\value{
Either \code{"qmd"} for package data or \code{"zip"} for uploaded data.
}
\description{
Return the download extension for an analysis recipe
}
\keyword{internal}
