% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-variable-transformations.R
\name{arithmeticTransformationInfo}
\alias{arithmeticTransformationInfo}
\title{Build simple arithmetic transformation information}
\usage{
arithmeticTransformationInfo(funName, constant, variablePosition)
}
\arguments{
\item{funName}{Character arithmetic operator.}

\item{constant}{Numeric constant.}

\item{variablePosition}{Character position of the data variable.}
}
\value{
Transformation information.
}
\description{
Build simple arithmetic transformation information
}
\keyword{internal}
