% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-explanationTeachingSummary.R
\name{backtickNames}
\alias{backtickNames}
\title{Wrap variable names in backticks for UI display}
\usage{
backtickNames(x)
}
\arguments{
\item{x}{Character vector.}
}
\value{
Character vector.
}
\description{
Wrap variable names in backticks for UI display
}
\keyword{internal}
