% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explain-badExplanation-types.R
\name{buildBadExplanationTypeGuide}
\alias{buildBadExplanationTypeGuide}
\title{Build textual definitions for supported bad explanation types}
\usage{
buildBadExplanationTypeGuide()
}
\value{
A character scalar.
}
\description{
Build textual definitions for supported bad explanation types
}
\keyword{internal}
