% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-ci-data.R
\name{buildConfidenceIntervalBaseInfo}
\alias{buildConfidenceIntervalBaseInfo}
\title{Build base-setting information for CI summaries}
\usage{
buildConfidenceIntervalBaseInfo(mf, numericReference, predictorNames)
}
\arguments{
\item{mf}{Model frame.}

\item{numericReference}{Numeric reference choice.}

\item{predictorNames}{Predictor names.}
}
\value{
A list with base-row information.
}
\description{
Build base-setting information for CI summaries
}
\keyword{internal}
