% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-ci-data.R
\name{buildConfidenceIntervalNotation}
\alias{buildConfidenceIntervalNotation}
\title{Build CI notation labels for supported models}
\usage{
buildConfidenceIntervalNotation(model, mf)
}
\arguments{
\item{model}{A fitted model object.}

\item{mf}{Model frame.}
}
\value{
A named list of notation labels.
}
\description{
Build CI notation labels for supported models
}
\keyword{internal}
