% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-ci-data.R
\name{buildFactorSettingGrid}
\alias{buildFactorSettingGrid}
\title{Build a grid of factor settings for teaching rows}
\usage{
buildFactorSettingGrid(mf, factorNames)
}
\arguments{
\item{mf}{Model frame.}

\item{factorNames}{Character vector of factor predictor names.}
}
\value{
A data frame of factor settings.
}
\description{
Build a grid of factor settings for teaching rows
}
\keyword{internal}
