% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-ci-data.R
\name{buildSimpleInteractionSlopeWeights}
\alias{buildSimpleInteractionSlopeWeights}
\title{Build coefficient weights for a simple interaction slope}
\usage{
buildSimpleInteractionSlopeWeights(model, factorName, factorLevel, numericName)
}
\arguments{
\item{model}{A fitted model object.}

\item{factorName}{Factor predictor name.}

\item{factorLevel}{Selected factor level.}

\item{numericName}{Numeric predictor name.}
}
\value{
A named numeric vector.
}
\description{
Build coefficient weights for a simple interaction slope
}
\keyword{internal}
