% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-plots.R
\name{chooseModelPlotPointAlpha}
\alias{chooseModelPlotPointAlpha}
\title{Choose a point alpha for model plots}
\usage{
chooseModelPlotPointAlpha(observationCount)
}
\arguments{
\item{observationCount}{Number of plotted observations.}
}
\value{
A numeric alpha value between 0 and 1.
}
\description{
Choose a point alpha for model plots
}
\keyword{internal}
