% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-explanationTeachingSummary.R
\name{collapseExplanationTeachingVariables}
\alias{collapseExplanationTeachingVariables}
\title{Collapse variable names into a readable list}
\usage{
collapseExplanationTeachingVariables(x)
}
\arguments{
\item{x}{Character vector of variable names.}
}
\value{
A single character string.
}
\description{
Collapse variable names into a readable list
}
\keyword{internal}
