% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-explanationTeachingSummary.R
\name{collapseTeachingNames}
\alias{collapseTeachingNames}
\title{Build a natural language list from names}
\usage{
collapseTeachingNames(x)
}
\arguments{
\item{x}{Character vector.}
}
\value{
A single character string.
}
\description{
Build a natural language list from names
}
\keyword{internal}
