% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-ci-data.R
\name{computeModelCriticalValue}
\alias{computeModelCriticalValue}
\title{Compute the critical value used in model confidence intervals}
\usage{
computeModelCriticalValue(model, level)
}
\arguments{
\item{model}{A fitted model object.}

\item{level}{Confidence level.}
}
\value{
A single numeric critical value.
}
\description{
Compute the critical value used in model confidence intervals
}
\keyword{internal}
