% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrastsUiHelpers.R
\name{contrastsAvgDragUi}
\alias{contrastsAvgDragUi}
\title{Drag-and-drop average contrast UI}
\usage{
contrastsAvgDragUi(levs)
}
\arguments{
\item{levs}{Factor levels.}
}
\description{
The user drags levels from the pool into left and right buckets. The contrast
is mean(left) - mean(right).
}
