% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrastsUiHelpers.R
\name{contrastsPairsUi}
\alias{contrastsPairsUi}
\title{Contrasts UI helper functions}
\usage{
contrastsPairsUi(levs)
}
\description{
These functions create the UI pieces used by the contrasts module.
}
