% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-server-reactive-state.R
\name{createAppServerReactiveState}
\alias{createAppServerReactiveState}
\title{Create app server reactive state}
\usage{
createAppServerReactiveState()
}
\value{
A list containing the app server reactive values and reactive holders.
}
\description{
Create app server reactive state
}
\keyword{internal}
