% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-variable-transformations.R
\name{deparseOneLine}
\alias{deparseOneLine}
\title{Deparse an expression to one line}
\usage{
deparseOneLine(expr)
}
\arguments{
\item{expr}{Expression to deparse.}
}
\value{
Character scalar.
}
\description{
Deparse an expression to one line
}
\keyword{internal}
