% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-explanationClaimEvidence.R
\name{detectComparison}
\alias{detectComparison}
\title{Detect comparison wording in a sentence}
\usage{
detectComparison(claimText, audit)
}
\arguments{
\item{claimText}{Character scalar.}

\item{audit}{A \code{wmfmExplanationAudit} object.}
}
\value{
Logical scalar.
}
\description{
Detect comparison wording in a sentence
}
\keyword{internal}
