% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-developer-scoring-grading.R
\name{detectDeveloperScoringFactorPredictors}
\alias{detectDeveloperScoringFactorPredictors}
\title{Detect whether a model contains factor predictors}
\usage{
detectDeveloperScoringFactorPredictors(model)
}
\arguments{
\item{model}{A fitted model object.}
}
\value{
Logical scalar.
}
\description{
Detect whether a model contains factor predictors
}
\keyword{internal}
