% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-developer-scoring-grading.R
\name{detectDeveloperScoringInteractionTerms}
\alias{detectDeveloperScoringInteractionTerms}
\title{Detect whether a model contains factor predictors}
\usage{
detectDeveloperScoringInteractionTerms(model)
}
\arguments{
\item{model}{A fitted model object.}
}
\value{
Logical scalar.

Logical scalar.
}
\description{
Detect whether a model contains factor predictors
}
\keyword{Detect}
\keyword{a}
\keyword{developer}
\keyword{has}
\keyword{interaction}
\keyword{internal}
\keyword{model}
\keyword{scoring}
\keyword{terms}
\keyword{whether}
