% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-explanationClaimEvidence.R
\name{detectEffect}
\alias{detectEffect}
\title{Detect effect wording in a sentence}
\usage{
detectEffect(claimText, audit, model = NULL)
}
\arguments{
\item{claimText}{Character scalar.}

\item{audit}{A \code{wmfmExplanationAudit} object.}

\item{model}{Optional fitted model object.}
}
\value{
Logical scalar.
}
\description{
Detect effect wording in a sentence
}
\keyword{internal}
