% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-explanationClaimEvidence.R
\name{detectUncertainty}
\alias{detectUncertainty}
\title{Detect uncertainty wording in a sentence}
\usage{
detectUncertainty(claimText)
}
\arguments{
\item{claimText}{Character scalar.}
}
\value{
Logical scalar.
}
\description{
Detect uncertainty wording in a sentence
}
\keyword{internal}
