% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text-detect-pattern.R
\name{detectWmfmPattern}
\alias{detectWmfmPattern}
\title{Detect pattern in text}
\usage{
detectWmfmPattern(x, pattern)
}
\arguments{
\item{x}{Character text.}

\item{pattern}{Regex pattern.}
}
\value{
Logical.
}
\description{
Detect pattern in text
}
