% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-provider-config.R
\name{editWmfmConfig}
\alias{editWmfmConfig}
\title{Edit the WMFM local configuration file}
\usage{
editWmfmConfig(editor = utils::file.edit)
}
\arguments{
\item{editor}{Function used to open the configuration file. Defaults to
\code{\link[utils:file.edit]{utils::file.edit()}}. This argument is mainly provided so tests can avoid
opening an editor.}
}
\value{
Invisibly returns the path to the configuration file.
}
\description{
Opens the WMFM user configuration file in the default R editor. If the
configuration directory or file does not yet exist, they are created first.
}
