% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-developer-scoring-grading.R
\name{extractDeveloperScoringInteractionInfo}
\alias{extractDeveloperScoringInteractionInfo}
\title{Extract developer scoring interaction metadata}
\usage{
extractDeveloperScoringInteractionInfo(model)
}
\arguments{
\item{model}{A fitted model object.}
}
\value{
A list with interaction term metadata.
}
\description{
Extract developer scoring interaction metadata
}
\keyword{internal}
