% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-cli-progress.R
\name{finishWmfmCliStageTracker}
\alias{finishWmfmCliStageTracker}
\title{Finish a simple command-line stage tracker}
\usage{
finishWmfmCliStageTracker(cliTracker)
}
\arguments{
\item{cliTracker}{A tracker created by \code{newWmfmCliStageTracker()}.}
}
\value{
Invisibly returns a list with total timing information.
}
\description{
Finish a simple command-line stage tracker
}
\keyword{internal}
