% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-response-backtransformation.R
\name{firstSourceVariable}
\alias{firstSourceVariable}
\title{Return the first source variable in a transformation record}
\usage{
firstSourceVariable(record)
}
\arguments{
\item{record}{A transformation record.}
}
\value{
Character scalar.
}
\description{
Return the first source variable in a transformation record
}
\keyword{internal}
