% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatBadExplanationOutput.R
\name{formatBadExplanationOutput}
\alias{formatBadExplanationOutput}
\title{Format bad explanation output for user-facing return}
\usage{
formatBadExplanationOutput(parsed, labelErrors = FALSE)
}
\arguments{
\item{parsed}{A validated parsed response.}

\item{labelErrors}{Logical. Should metadata be returned?}
}
\value{
A character scalar, a named character vector, or a named list.
}
\description{
Format bad explanation output for user-facing return
}
\keyword{internal}
