% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-ci-data.R
\name{formatConfidenceIntervalNumber}
\alias{formatConfidenceIntervalNumber}
\title{Format a number for CI display text}
\usage{
formatConfidenceIntervalNumber(x)
}
\arguments{
\item{x}{A numeric value.}
}
\value{
A character scalar.
}
\description{
Format a number for CI display text
}
\keyword{internal}
