% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-explanationFormat.R
\name{formatExplanationMultiplier}
\alias{formatExplanationMultiplier}
\title{Format a multiplicative change for student-facing explanation text}
\usage{
formatExplanationMultiplier(multiplier, sigDigits = 2)
}
\arguments{
\item{multiplier}{Numeric vector.}

\item{sigDigits}{Number of significant figures to keep. Defaults to 2.}
}
\value{
A character vector.
}
\description{
Format a multiplicative change for student-facing explanation text
}
\keyword{internal}
