% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-formula-derived-response.R
\name{formulaExpressionsMatch}
\alias{formulaExpressionsMatch}
\title{Test whether two formula expressions are textually equivalent}
\usage{
formulaExpressionsMatch(first, second)
}
\arguments{
\item{first}{Character scalar expression text.}

\item{second}{Character scalar expression text.}
}
\value{
Logical scalar.
}
\description{
Test whether two formula expressions are textually equivalent
}
\keyword{internal}
