% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-adjustment-variables.R
\name{getAdjustmentCheckboxOnChangeScript}
\alias{getAdjustmentCheckboxOnChangeScript}
\title{Build onchange JavaScript for inline adjustment checkboxes}
\usage{
getAdjustmentCheckboxOnChangeScript()
}
\value{
Character scalar JavaScript snippet.
}
\description{
Build onchange JavaScript for inline adjustment checkboxes
}
\keyword{internal}
