% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explain-badExplanation-types.R
\name{getAvailableBadExplanationTypes}
\alias{getAvailableBadExplanationTypes}
\title{Get bad explanation types compatible with a WMFM model}
\usage{
getAvailableBadExplanationTypes(x)
}
\arguments{
\item{x}{A \code{wmfmModel} object.}
}
\value{
A character vector of compatible bad explanation types.
}
\description{
Get bad explanation types compatible with a WMFM model
}
\keyword{internal}
