% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-ci-data.R
\name{getConfidenceIntervalCriticalValue}
\alias{getConfidenceIntervalCriticalValue}
\title{Get the CI critical value for a fitted model}
\usage{
getConfidenceIntervalCriticalValue(model, level)
}
\arguments{
\item{model}{A fitted model object.}

\item{level}{Confidence level.}
}
\value{
A numeric scalar.
}
\description{
Get the CI critical value for a fitted model
}
\keyword{internal}
