% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples-installedPackagesWithData.R
\name{getInstalledPackagesWithData}
\alias{getInstalledPackagesWithData}
\title{List installed packages that contain datasets}
\usage{
getInstalledPackagesWithData()
}
\value{
A character vector of installed package names that appear to
contain one or more datasets.
}
\description{
Returns the names of installed packages for which
\code{utils::data(package = ...)} reports at least one dataset.
The result is sorted alphabetically and is useful for populating
a package selector in the app's Load Data tab.
}
\keyword{internal}
