% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-variable-transformations.R
\name{getModelResponseTransformationMode}
\alias{getModelResponseTransformationMode}
\title{Get response-transformation handling metadata from a fitted model}
\usage{
getModelResponseTransformationMode(model)
}
\arguments{
\item{model}{A fitted model object.}
}
\value{
One of \code{"both"}, \code{"model"}, or \code{"original"}.
}
\description{
Get response-transformation handling metadata from a fitted model
}
\keyword{internal}
