% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-response-backtransformation.R
\name{getModelResponseTransformationRecord}
\alias{getModelResponseTransformationRecord}
\title{Get the fitted-model response transformation record}
\usage{
getModelResponseTransformationRecord(model)
}
\arguments{
\item{model}{A fitted model object.}
}
\value{
A transformation record or NULL.
}
\description{
Get the fitted-model response transformation record
}
\keyword{internal}
