% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-variable-transformations.R
\name{getModelVariableTransformationRoles}
\alias{getModelVariableTransformationRoles}
\title{Get transformation roles for fitted model variables}
\usage{
getModelVariableTransformationRoles(model)
}
\arguments{
\item{model}{A fitted model object.}
}
\value{
A named character vector with values \code{response}, \code{predictor}, or
\code{unused}.
}
\description{
Get transformation roles for fitted model variables
}
\keyword{internal}
