% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-packageData.R
\name{getPackageNameFromDescription}
\alias{getPackageNameFromDescription}
\title{Get a package name from an installed package DESCRIPTION file}
\usage{
getPackageNameFromDescription(packageDir)
}
\arguments{
\item{packageDir}{Path to an installed package directory.}
}
\value{
A character scalar package name, or an empty string if the name
cannot be read.
}
\description{
Get a package name from an installed package DESCRIPTION file
}
\keyword{internal}
