% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-packageData.R
\name{getRdAliases}
\alias{getRdAliases}
\title{Extract aliases from an Rd object}
\usage{
getRdAliases(rd)
}
\arguments{
\item{rd}{An Rd object.}
}
\value{
Character vector of aliases found in the Rd object.
}
\description{
Extract aliases from an Rd object
}
\keyword{internal}
