% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getS20xDocHtml.R
\name{getS20xDocHtml}
\alias{getS20xDocHtml}
\title{Get s20x dataset documentation as HTML}
\usage{
getS20xDocHtml(dataset)
}
\arguments{
\item{dataset}{Character string, name of a data set in the s20x package.}
}
\value{
A single character string containing HTML for the help page, or NULL if not found.
}
\description{
Get s20x dataset documentation as HTML
}
\keyword{internal}
