% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getS20xDocText.R
\name{getS20xDocText}
\alias{getS20xDocText}
\title{Get s20x dataset documentation as plain text}
\usage{
getS20xDocText(dataset)
}
\arguments{
\item{dataset}{Character string, name of a data set in the s20x package.}
}
\value{
A single character string with the Rd help text, or NULL if not found.
}
\description{
Get s20x dataset documentation as plain text
}
\keyword{internal}
