% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-provider-config.R
\name{getWmfmProviderAdapter}
\alias{getWmfmProviderAdapter}
\title{Get WMFM provider adapter metadata}
\usage{
getWmfmProviderAdapter(provider)
}
\arguments{
\item{provider}{Character scalar provider id.}
}
\value{
Named list with provider adapter metadata.
}
\description{
Get WMFM provider adapter metadata
}
\keyword{internal}
