% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWmfmRunsClaimsData.R
\name{getWmfmRunsClaimsData}
\alias{getWmfmRunsClaimsData}
\title{Build extracted-claim frequency data for a WMFM runs object}
\usage{
getWmfmRunsClaimsData(x)
}
\arguments{
\item{x}{A \code{wmfmRuns} object.}
}
\value{
A data frame summarising claim frequencies.
}
\description{
Build extracted-claim frequency data for a WMFM runs object
}
