% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-server-data-load-helpers.R
\name{hasPackageDatasetContext}
\alias{hasPackageDatasetContext}
\title{Determine whether a package data set has built-in context.}
\usage{
hasPackageDatasetContext(dataSource, packageName, datasetName)
}
\arguments{
\item{dataSource}{Character scalar identifying the active data source.}

\item{packageName}{Character scalar package name.}

\item{datasetName}{Character scalar data set name.}
}
\value{
\code{TRUE} when WMFM can use package documentation as data context.
}
\description{
Determine whether a package data set has built-in context.
}
\keyword{internal}
