% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-variable-transformations.R
\name{inferArithmeticTransformation}
\alias{inferArithmeticTransformation}
\title{Infer a simple arithmetic transformation label}
\usage{
inferArithmeticTransformation(rhs, funName)
}
\arguments{
\item{rhs}{Parsed right-hand-side expression.}

\item{funName}{Character arithmetic operator.}
}
\value{
Transformation information or \code{NULL}.
}
\description{
Infer a simple arithmetic transformation label
}
\keyword{internal}
