% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-variable-transformations.R
\name{inverseTypeForFunction}
\alias{inverseTypeForFunction}
\title{Get inverse label for a recognised transformation function}
\usage{
inverseTypeForFunction(funName)
}
\arguments{
\item{funName}{Character function name.}
}
\value{
Character inverse label.
}
\description{
Get inverse label for a recognised transformation function
}
\keyword{internal}
