% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-analysisRecipe.R
\name{isAnalysisRecipe}
\alias{isAnalysisRecipe}
\title{Test whether an object is a WMFM analysis recipe}
\usage{
isAnalysisRecipe(x)
}
\arguments{
\item{x}{Object to test.}
}
\value{
\code{TRUE} when \code{x} is a valid classed analysis recipe.
}
\description{
Test whether an object is a WMFM analysis recipe
}
\keyword{internal}
