% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-ci-data.R
\name{isLmIdentityModel}
\alias{isLmIdentityModel}
\title{Check whether a model is a linear identity-scale model}
\usage{
isLmIdentityModel(model)
}
\arguments{
\item{model}{A fitted model object.}
}
\value{
A logical scalar.
}
\description{
Check whether a model is a linear identity-scale model
}
\keyword{internal}
