% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-ci-data.R
\name{isSupportedLogisticModel}
\alias{isSupportedLogisticModel}
\title{Test whether a model is a supported logistic GLM}
\usage{
isSupportedLogisticModel(model)
}
\arguments{
\item{model}{A fitted model object.}
}
\value{
A logical scalar.
}
\description{
Test whether a model is a supported logistic GLM
}
\keyword{internal}
