% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-ci-data.R
\name{isSupportedPoissonModel}
\alias{isSupportedPoissonModel}
\title{Test whether a model is a supported Poisson GLM}
\usage{
isSupportedPoissonModel(model)
}
\arguments{
\item{model}{A fitted model object.}
}
\value{
A logical scalar.
}
\description{
Test whether a model is a supported Poisson GLM
}
\keyword{internal}
