% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-response-backtransformation.R
\name{isUsableResponseInverseType}
\alias{isUsableResponseInverseType}
\title{Check whether an inverse type can be used for response back-transformation}
\usage{
isUsableResponseInverseType(inverseType)
}
\arguments{
\item{inverseType}{Character inverse label.}
}
\value{
Logical scalar.
}
\description{
Check whether an inverse type can be used for response back-transformation
}
\keyword{internal}
