% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explain-badExplanation-listTypes.R
\name{listBadExplanationTypes}
\alias{listBadExplanationTypes}
\title{List supported bad explanation types}
\usage{
listBadExplanationTypes()
}
\value{
A character vector of supported bad explanation types.
}
\description{
Returns the currently supported error types that can be requested when
generating deliberately flawed model explanations.
}
