% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-provider-config.R
\name{listWmfmProviderAdapters}
\alias{listWmfmProviderAdapters}
\title{List WMFM provider adapters}
\usage{
listWmfmProviderAdapters()
}
\value{
Character vector of registry provider ids.
}
\description{
List WMFM provider adapters
}
\keyword{internal}
