% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-cli-progress.R
\name{newWmfmCliStageTracker}
\alias{newWmfmCliStageTracker}
\title{Create a simple command-line stage tracker}
\usage{
newWmfmCliStageTracker(
  showProgress = TRUE,
  showTiming = TRUE,
  taskLabel = "Task"
)
}
\arguments{
\item{showProgress}{Logical. Should progress messages be shown?}

\item{showTiming}{Logical. Should timing summaries be shown?}

\item{taskLabel}{Character scalar naming the task.}
}
\value{
A list used internally for command-line progress and timing.
}
\description{
Create a simple command-line stage tracker
}
\keyword{internal}
