% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text-utils.R
\name{normaliseWmfmText}
\alias{normaliseWmfmText}
\title{Normalise text for comparison}
\usage{
normaliseWmfmText(x)
}
\arguments{
\item{x}{Character text.}
}
\value{
Character string.
}
\description{
Lowercases and removes extra whitespace.
}
