% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizeWmfmText.R
\name{normalizeWmfmText}
\alias{normalizeWmfmText}
\title{Normalize text for comparison}
\usage{
normalizeWmfmText(x)
}
\arguments{
\item{x}{Character text.}
}
\value{
Character string.
}
\description{
Lowercases and removes extra whitespace.
}
