% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.wmfmScoreComparison.R
\name{plot.wmfmScoreComparison}
\alias{plot.wmfmScoreComparison}
\title{Plot a WMFM score comparison object}
\usage{
\method{plot}{wmfmScoreComparison}(x, type = c("agreement", "overall"), ...)
}
\arguments{
\item{x}{A \code{wmfmScoreComparison} object.}

\item{type}{"agreement" or "overall"}
}
\description{
Plot a WMFM score comparison object
}
