% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-explanation-cleanText.R
\name{postProcessConfidenceIntervalTerminology}
\alias{postProcessConfidenceIntervalTerminology}
\title{Standardise recurring confidence interval terminology}
\usage{
postProcessConfidenceIntervalTerminology(text)
}
\arguments{
\item{text}{Character vector.}
}
\value{
A character vector with safer confidence interval terminology.
}
\description{
Standardise recurring confidence interval terminology
}
\keyword{internal}
