% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-explanation-cleanText.R
\name{postProcessExtractNumericTokens}
\alias{postProcessExtractNumericTokens}
\title{Extract numeric tokens from explanation text}
\usage{
postProcessExtractNumericTokens(text)
}
\arguments{
\item{text}{Character scalar.}
}
\value{
A character vector of numeric tokens.
}
\description{
Extract numeric tokens from explanation text
}
\keyword{internal}
