% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-explanation-cleanText.R
\name{postProcessGrammarCleanup}
\alias{postProcessGrammarCleanup}
\title{Clean small grammar artefacts created by deterministic rewrites}
\usage{
postProcessGrammarCleanup(text)
}
\arguments{
\item{text}{Character vector.}
}
\value{
A character vector with small grammar artefacts cleaned up.
}
\description{
Clean small grammar artefacts created by deterministic rewrites
}
\keyword{internal}
