% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-explanation-cleanText.R
\name{postProcessIssueDataFrame}
\alias{postProcessIssueDataFrame}
\title{Create an empty surface issue data frame}
\usage{
postProcessIssueDataFrame()
}
\value{
An empty data frame using the surface issue schema.
}
\description{
Create an empty surface issue data frame
}
\keyword{internal}
