% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-explanation-cleanText.R
\name{postProcessModelMechanismLanguage}
\alias{postProcessModelMechanismLanguage}
\title{Remove a small set of model-mechanism phrasings}
\usage{
postProcessModelMechanismLanguage(text)
}
\arguments{
\item{text}{Character vector.}
}
\value{
A character vector with recurring technical leakage reduced.
}
\description{
Remove a small set of model-mechanism phrasings
}
\keyword{internal}
