% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-explanation-cleanText.R
\name{postProcessSurfaceIssueRules}
\alias{postProcessSurfaceIssueRules}
\title{Surface issue scan rules}
\usage{
postProcessSurfaceIssueRules()
}
\value{
A list of diagnostic surface issue rules.
}
\description{
Surface issue scan rules
}
\keyword{internal}
