% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-explanation-cleanText.R
\name{postProcessVerbalFractions}
\alias{postProcessVerbalFractions}
\title{Replace verbal fractions in numeric contexts}
\usage{
postProcessVerbalFractions(text)
}
\arguments{
\item{text}{Character vector.}
}
\value{
A character vector with verbal fractions replaced by percentages.
}
\description{
Replace verbal fractions in numeric contexts
}
\keyword{internal}
