% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-print-metricComparisonSummary.R
\name{print.metricComparisonSummary}
\alias{print.metricComparisonSummary}
\title{Print a metric comparison summary}
\usage{
\method{print}{metricComparisonSummary}(x, ...)
}
\arguments{
\item{x}{A \code{metricComparisonSummary} object.}

\item{...}{Unused. Included for S3 compatibility.}
}
\value{
The input object, invisibly.
}
\description{
Print a metric comparison summary
}
