% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-print-summary-wmfmGrade.R
\name{print.summary.wmfmGrade}
\alias{print.summary.wmfmGrade}
\title{Print a WMFM grade summary}
\usage{
\method{print}{summary.wmfmGrade}(x, digits = 2, ...)
}
\arguments{
\item{x}{A \code{summary.wmfmGrade} object.}

\item{digits}{Number of digits for numeric output.}

\item{...}{Unused.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a WMFM grade summary
}
