% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-print-summary-wmfmGradeListObj.R
\name{print.summary.wmfmGradeListObj}
\alias{print.summary.wmfmGradeListObj}
\title{Print a WMFM grade list summary}
\usage{
\method{print}{summary.wmfmGradeListObj}(x, digits = 2, ...)
}
\arguments{
\item{x}{A \code{summary.wmfmGradeListObj} object.}

\item{digits}{Number of digits for numeric output.}

\item{...}{Unused.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a WMFM grade list summary
}
