% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-print-summary-wmfmRuns.R
\name{print.summary.wmfmRuns}
\alias{print.summary.wmfmRuns}
\title{Print a summary.wmfmRuns object}
\usage{
\method{print}{summary.wmfmRuns}(x, ...)
}
\arguments{
\item{x}{A \code{summary.wmfmRuns} object.}

\item{...}{Reserved for future extensions.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a summary.wmfmRuns object
}
