% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-explanation-cleanText.R
\name{print.wmfmExplanationSurfaceDiagnosis}
\alias{print.wmfmExplanationSurfaceDiagnosis}
\title{Print an explanation surface diagnosis}
\usage{
\method{print}{wmfmExplanationSurfaceDiagnosis}(x, ...)
}
\arguments{
\item{x}{A \code{wmfmExplanationSurfaceDiagnosis} object.}

\item{...}{Additional arguments, currently unused.}
}
\value{
The input object, invisibly.
}
\description{
Print an explanation surface diagnosis
}
