% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.wmfmGradeComparison.R
\name{print.wmfmGradeComparison}
\alias{print.wmfmGradeComparison}
\title{Print a wmfmGradeComparison object}
\usage{
\method{print}{wmfmGradeComparison}(x, digits = 2, maxRows = 6, ...)
}
\arguments{
\item{x}{A \code{wmfmGradeComparison} object.}

\item{digits}{Number of digits for numeric output.}

\item{maxRows}{Maximum number of rows to print in each section.}

\item{...}{Unused. Included for S3 compatibility.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a wmfmGradeComparison object
}
