% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-print-wmfmGradeListObj.R
\name{print.wmfmGradeListObj}
\alias{print.wmfmGradeListObj}
\title{Print a WMFM grade list object}
\usage{
\method{print}{wmfmGradeListObj}(x, digits = 2, maxExamples = 6, ...)
}
\arguments{
\item{x}{A \code{wmfmGradeListObj} object.}

\item{digits}{Number of digits for numeric output.}

\item{maxExamples}{Maximum number of explanation marks to print.}

\item{...}{Unused.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a WMFM grade list object
}
