% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.wmfmMetricDiagnosis.R
\name{print.wmfmMetricDiagnosis}
\alias{print.wmfmMetricDiagnosis}
\title{Print a metric diagnosis object}
\usage{
\method{print}{wmfmMetricDiagnosis}(x, ...)
}
\arguments{
\item{x}{A \code{wmfmMetricDiagnosis} object.}

\item{...}{Unused.}
}
\value{
The input object, invisibly.
}
\description{
Print a metric diagnosis object
}
