% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-print-wmfmScoresDiagnosis.R
\name{print.wmfmScoresDiagnosis}
\alias{print.wmfmScoresDiagnosis}
\title{Print a scores diagnosis object}
\usage{
\method{print}{wmfmScoresDiagnosis}(x, ...)
}
\arguments{
\item{x}{A \code{wmfmScoresDiagnosis} object.}

\item{...}{Unused.}
}
\value{
The input object, invisibly.
}
\description{
Print a scores diagnosis object
}
