% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-provider-config.R
\name{readWmfmConfigPath}
\alias{readWmfmConfigPath}
\title{Read the WMFM local configuration file path}
\usage{
readWmfmConfigPath()
}
\value{
Character scalar path to \code{config.json}.
}
\description{
Alias for \code{\link[=getWmfmConfigPath]{getWmfmConfigPath()}}.
}
