% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-server-analysis-download.R
\name{registerAnalysisDownloadObserver}
\alias{registerAnalysisDownloadObserver}
\title{Register the reproducible-analysis download handler}
\usage{
registerAnalysisDownloadObserver(output, rv)
}
\arguments{
\item{output}{Shiny output object.}

\item{rv}{App reactive values object.}
}
\value{
No return value; called for observer side effects.
}
\description{
Register the reproducible-analysis download handler
}
\keyword{internal}
