% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registerContrastsModule.R
\name{registerContrastsModule}
\alias{registerContrastsModule}
\title{Register the contrasts module}
\usage{
registerContrastsModule(input, output, session, rv, modelFit, modelFrame)
}
\arguments{
\item{input, output, session}{Shiny server objects}

\item{rv}{reactiveValues() list (must contain chatProvider, contrastLlmCache)}

\item{modelFit}{reactive returning fitted model}

\item{modelFrame}{reactive returning model.frame(model)}
}
\description{
Register the contrasts module
}
