% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-server-explanation.R
\name{registerModelExplanationObservers}
\alias{registerModelExplanationObservers}
\title{Register model explanation observers for the app server}
\usage{
registerModelExplanationObservers(
  input,
  output,
  rv,
  modelFit,
  developerModeUnlocked
)
}
\arguments{
\item{input}{Shiny input object.}

\item{output}{Shiny output object.}

\item{rv}{App reactive values object.}

\item{modelFit}{Reactive value containing the fitted model.}

\item{developerModeUnlocked}{Reactive value indicating developer mode status.}
}
\value{
A list of model-explanation reactive helpers.
}
\description{
Register model explanation observers for the app server
}
\keyword{internal}
