% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-server-model-output.R
\name{registerModelSummaryObservers}
\alias{registerModelSummaryObservers}
\title{Register model output observers for the app server}
\usage{
registerModelSummaryObservers(output, modelFit)
}
\arguments{
\item{output}{Shiny output object.}

\item{modelFit}{Reactive value containing the fitted model.}
}
\value{
No return value; called for output side effects.
}
\description{
Register model output observers for the app server
}
\keyword{internal}
