% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis-recipe-render-analysis.R
\name{renderAnalysisRecipeAnalysisQuarto}
\alias{renderAnalysisRecipeAnalysisQuarto}
\title{Render post-fit analysis sections for a WMFM analysis recipe}
\usage{
renderAnalysisRecipeAnalysisQuarto(recipe)
}
\arguments{
\item{recipe}{A \code{wmfmAnalysisRecipe} object.}
}
\value{
Character vector containing Quarto source lines.
}
\description{
Render post-fit analysis sections for a WMFM analysis recipe
}
\keyword{internal}
