% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis-recipe-render-analysis.R
\name{renderAnalysisRecipeAnovaChunk}
\alias{renderAnalysisRecipeAnovaChunk}
\title{Render analysis-of-variance code}
\usage{
renderAnalysisRecipeAnovaChunk(recipe)
}
\arguments{
\item{recipe}{A \code{wmfmAnalysisRecipe} object.}
}
\value{
Character vector containing a Quarto code chunk.
}
\description{
Render analysis-of-variance code
}
\keyword{internal}
