% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis-recipe-render-analysis.R
\name{renderAnalysisRecipeConfidenceIntervalChunk}
\alias{renderAnalysisRecipeConfidenceIntervalChunk}
\title{Render confidence-interval code}
\usage{
renderAnalysisRecipeConfidenceIntervalChunk(recipe)
}
\arguments{
\item{recipe}{A \code{wmfmAnalysisRecipe} object.}
}
\value{
Character vector containing a Quarto code chunk.
}
\description{
Render confidence-interval code
}
\keyword{internal}
