% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis-recipe-render-core.R
\name{renderAnalysisRecipeDataChunk}
\alias{renderAnalysisRecipeDataChunk}
\title{Render the data-loading code chunk}
\usage{
renderAnalysisRecipeDataChunk(recipe)
}
\arguments{
\item{recipe}{A \code{wmfmAnalysisRecipe} object.}
}
\value{
Character vector containing a Quarto code chunk.
}
\description{
Render the data-loading code chunk
}
\keyword{internal}
