% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis-recipe-render-analysis.R
\name{renderAnalysisRecipeDiagnosticChunk}
\alias{renderAnalysisRecipeDiagnosticChunk}
\title{Render diagnostic-plot code}
\usage{
renderAnalysisRecipeDiagnosticChunk(recipe)
}
\arguments{
\item{recipe}{A \code{wmfmAnalysisRecipe} object.}
}
\value{
Character vector containing a Quarto code chunk.
}
\description{
Render diagnostic-plot code
}
\keyword{internal}
