% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis-recipe-render-core.R
\name{renderAnalysisRecipeModelChunk}
\alias{renderAnalysisRecipeModelChunk}
\title{Render the model-fitting code chunk}
\usage{
renderAnalysisRecipeModelChunk(recipe)
}
\arguments{
\item{recipe}{A \code{wmfmAnalysisRecipe} object.}
}
\value{
Character vector containing a Quarto code chunk.
}
\description{
Render the model-fitting code chunk
}
\keyword{internal}
