% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis-recipe-render-analysis.R
\name{renderAnalysisRecipeModelPlotChunk}
\alias{renderAnalysisRecipeModelPlotChunk}
\title{Render the substantive model-plot code}
\usage{
renderAnalysisRecipeModelPlotChunk(recipe)
}
\arguments{
\item{recipe}{A \code{wmfmAnalysisRecipe} object.}
}
\value{
Character vector containing a Quarto code chunk.
}
\description{
Render the substantive model-plot code
}
\keyword{internal}
