% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis-recipe-render-core.R
\name{renderAnalysisRecipePackageChunk}
\alias{renderAnalysisRecipePackageChunk}
\title{Render the package-loading code chunk}
\usage{
renderAnalysisRecipePackageChunk(recipe)
}
\arguments{
\item{recipe}{A \code{wmfmAnalysisRecipe} object.}
}
\value{
Character vector containing a Quarto code chunk.
}
\description{
Render the package-loading code chunk
}
\keyword{internal}
