% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis-recipe-render-core.R
\name{renderAnalysisRecipePreparationChunk}
\alias{renderAnalysisRecipePreparationChunk}
\title{Render the data-preparation code chunk}
\usage{
renderAnalysisRecipePreparationChunk(recipe)
}
\arguments{
\item{recipe}{A \code{wmfmAnalysisRecipe} object.}
}
\value{
Character vector containing a Quarto code chunk.
}
\description{
Render the data-preparation code chunk
}
\keyword{internal}
