% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis-recipe-render-core.R
\name{renderAnalysisRecipePreparationSection}
\alias{renderAnalysisRecipePreparationSection}
\title{Render the data-preparation section}
\usage{
renderAnalysisRecipePreparationSection(recipe)
}
\arguments{
\item{recipe}{A \code{wmfmAnalysisRecipe} object.}
}
\value{
Character vector containing explanatory text and a Quarto code chunk.
}
\description{
Render the data-preparation section
}
\keyword{internal}
