% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis-recipe-render-analysis.R
\name{renderAnalysisRecipeSummaryChunk}
\alias{renderAnalysisRecipeSummaryChunk}
\title{Render model-summary code}
\usage{
renderAnalysisRecipeSummaryChunk(recipe)
}
\arguments{
\item{recipe}{A \code{wmfmAnalysisRecipe} object.}
}
\value{
Character vector containing a Quarto code chunk.
}
\description{
Render model-summary code
}
\keyword{internal}
