% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-adjustment-variables.R
\name{renderBucketVariableLabel}
\alias{renderBucketVariableLabel}
\title{Build a variable-card label with inline adjustment checkbox}
\usage{
renderBucketVariableLabel(variableName, selectedAdjustmentVariables)
}
\arguments{
\item{variableName}{Character scalar variable name to render.}

\item{selectedAdjustmentVariables}{Character vector of currently selected
adjustment variables.}
}
\value{
A Shiny tag object suitable for bucket list labels.
}
\description{
Build a variable-card label with inline adjustment checkbox
}
\keyword{internal}
