% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolveResponseNounPhrase.R
\name{resolveResponseNounPhrase}
\alias{resolveResponseNounPhrase}
\title{Resolve the response noun phrase for a fitted model}
\usage{
resolveResponseNounPhrase(model, responseVar)
}
\arguments{
\item{model}{Fitted model object.}

\item{responseVar}{Character scalar response variable name.}
}
\value{
Character scalar noun phrase.
}
\description{
Resolve the response noun phrase for a fitted model
}
\keyword{internal}
