% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{score}
\alias{score}
\title{Score WMFM objects}
\usage{
score(x, ...)
}
\arguments{
\item{x}{An object to score.}

\item{...}{Additional arguments passed to methods.}
}
\value{
Method-specific scored output.
}
\description{
Generic for scoring WMFM objects.
}
