% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-explanationClaimEvidence.R
\name{sentenceMatchesBaselineEvidence}
\alias{sentenceMatchesBaselineEvidence}
\title{Check whether a sentence matches baseline evidence}
\usage{
sentenceMatchesBaselineEvidence(claimText, audit)
}
\arguments{
\item{claimText}{Character scalar.}

\item{audit}{A \code{wmfmExplanationAudit} object.}
}
\value{
Logical scalar.
}
\description{
Check whether a sentence matches baseline evidence
}
\keyword{internal}
