% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-stability.R
\name{stability}
\alias{stability}
\title{Assess stability of WMFM objects}
\usage{
stability(x, ...)
}
\arguments{
\item{x}{An object to assess.}

\item{...}{Additional arguments passed to methods.}
}
\value{
Method-specific stability output.
}
\description{
Generic for assessing stability of WMFM objects.
}
