% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-stability-wmfmScores.R
\name{stability.wmfmScores}
\alias{stability.wmfmScores}
\title{Assess stability of WMFM score results}
\usage{
\method{stability}{wmfmScores}(x, ...)
}
\arguments{
\item{x}{A \code{wmfmScores} object.}

\item{...}{Unused. Included for S3 compatibility.}
}
\value{
An object of class \code{wmfmScoreStability}.
}
\description{
Summarises within-method stability of scores stored in a \code{wmfmScores} object.
Stability is summarised separately for binary, ordinal, and continuous score
fields.
}
