% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.wmfmGradeListObj.R
\name{summary.wmfmGradeListObj}
\alias{summary.wmfmGradeListObj}
\title{Summarise a WMFM grade list object}
\usage{
\method{summary}{wmfmGradeListObj}(object, ...)
}
\arguments{
\item{object}{A \code{wmfmGradeListObj} object.}

\item{...}{Unused.}
}
\value{
An object of class \code{summary.wmfmGradeListObj}.
}
\description{
Summarise a WMFM grade list object
}
