% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-analysisRecipe.R
\name{updateAnalysisRecipeSection}
\alias{updateAnalysisRecipeSection}
\title{Replace one analysis recipe section}
\usage{
updateAnalysisRecipeSection(recipe, sectionName, section)
}
\arguments{
\item{recipe}{A \code{wmfmAnalysisRecipe} object.}

\item{sectionName}{Character scalar naming the section.}

\item{section}{Structured replacement section.}
}
\value{
An updated \code{wmfmAnalysisRecipe} object.
}
\description{
Replace one analysis recipe section
}
\keyword{internal}
