% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-analysisRecipe.R
\name{validateAnalysisRecipe}
\alias{validateAnalysisRecipe}
\title{Validate a WMFM analysis recipe}
\usage{
validateAnalysisRecipe(recipe)
}
\arguments{
\item{recipe}{Object to validate.}
}
\value{
The validated recipe, invisibly.
}
\description{
Validate a WMFM analysis recipe
}
\keyword{internal}
