% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validResponseVars.R
\name{validResponseVars}
\alias{validResponseVars}
\title{List valid response variables for a model type}
\usage{
validResponseVars(data, modelType)
}
\arguments{
\item{data}{A data frame.}

\item{modelType}{Model type passed to \code{validateResponseVar()}.}
}
\value{
A character vector of variable names.
}
\description{
Returns the names of variables in a data frame that are supported
as response variables for the specified model type.
}
\keyword{internal}
