% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-provider-config.R
\name{wmfmConfigPath}
\alias{wmfmConfigPath}
\title{WMFM local configuration file path}
\usage{
wmfmConfigPath()
}
\value{
Character scalar path to local configuration file.
}
\description{
WMFM local configuration file path
}
\keyword{internal}
