% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-compare.R
\name{compare}
\alias{compare}
\title{Compare WMFM objects}
\usage{
compare(x, y = NULL, ...)
}
\arguments{
\item{x}{An object to compare.}

\item{y}{Optional second object to compare.}

\item{...}{Additional arguments passed to methods.}
}
\value{
Method specific comparison object.
}
\description{
Generic for comparing WMFM objects.
}
