% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-developer-scoring-grading.R
\name{detectDeveloperScoringInteractionTerms}
\alias{detectDeveloperScoringInteractionTerms}
\title{Detect whether a developer scoring model has interaction terms}
\usage{
detectDeveloperScoringInteractionTerms(model)
}
\arguments{
\item{model}{A fitted model object.}
}
\value{
Logical scalar.
}
\description{
Detect whether a developer scoring model has interaction terms
}
\keyword{internal}
