% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cacheEnv.R
\docType{data}
\name{.env_cache}
\alias{.env_cache}
\title{Environment for caching LLM results}
\format{
An object of class \code{environment} of length 0.
}
\usage{
.env_cache
}
\description{
Internal environment used to cache equations and explanations for
previously fitted models.
}
\keyword{internal}
