% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-student-explanation.R
\name{formatStudentExplanationNumber}
\alias{formatStudentExplanationNumber}
\title{Format a number for the student explanation toolbar}
\usage{
formatStudentExplanationNumber(value)
}
\arguments{
\item{value}{Numeric vector.}
}
\value{
A character vector.
}
\description{
Format a number for the student explanation toolbar
}
\keyword{internal}
