% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-plots.R
\name{plotModelPlot}
\alias{plotModelPlot}
\title{Draw a student-facing model plot}
\usage{
plotModelPlot(model, plotType = c("observedFitted", "residualFitted"))
}
\arguments{
\item{model}{A fitted model object or \code{wmfmModel} object.}

\item{plotType}{Plot type.}
}
\value{
A \code{ggplot} object, or \code{NULL} for unsupported models.
}
\description{
Draw a student-facing model plot
}
