% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-score.R
\name{plotWmfmScoreAgreementSummary}
\alias{plotWmfmScoreAgreementSummary}
\title{Plot ordinal agreement summary for WMFM score comparison}
\usage{
plotWmfmScoreAgreementSummary(x, orderBy = c("worst", "registry"))
}
\arguments{
\item{x}{A wmfmScoreComparison object}

\item{orderBy}{"worst" or "registry"}
}
\value{
ggplot object
}
\description{
Plot ordinal agreement summary for WMFM score comparison
}
