% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-student-explanation.R
\name{registerStudentExplanationObservers}
\alias{registerStudentExplanationObservers}
\title{Register student explanation workspace observers}
\usage{
registerStudentExplanationObservers(input, output, session, modelFit)
}
\arguments{
\item{input}{Shiny input object.}

\item{output}{Shiny output object.}

\item{session}{Shiny session object.}

\item{modelFit}{Reactive fitted model holder.}
}
\value{
Invisible \code{NULL}.
}
\description{
Register student explanation workspace observers
}
\keyword{internal}
