% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summariseWmfmRepeatedRuns.R
\name{summariseWmfmRepeatedRuns}
\alias{summariseWmfmRepeatedRuns}
\title{Summarise repeated runs}
\usage{
summariseWmfmRepeatedRuns(runsDf)
}
\arguments{
\item{runsDf}{Data frame of runs.}
}
\value{
List summary.
}
\description{
Summarise repeated runs
}
